+++ article-commentaire.php
@@ -1,8 +1,8 @@
 <?php
     if(session_status() == PHP_SESSION_NONE){ session_start();}
 
-    require_once ('Article.php');
-    require_once ('Comment.php');
+    require_once ('_Class/Article.php');
+    require_once ('_Class/Comment.php');
 
 
     $article = new Article();
@@ -28,7 +28,7 @@ if(isset($_GET['commentaires'])) {
 <html lang="en">
 <head>
     <?php require_once ('_include/head.php') ?>
-    <script defer src="article-commentaire.js"></script>
+    <script defer src="_js/article-commentaire.js"></script>
     <title>Article</title>
 </head>
 <body>
