+++ articles-page.php
@@ -3,13 +3,13 @@
 <html lang="en">
 <head>
     <?php require_once '_include/head.php' ?>
-    <script src="articles-page.js" defer></script>
+    <script src="_js/articles-page.js" defer></script>
     <title>Document</title>
 </head>
 <body>
 
 
-    <?php require_once 'Article.php' ?>
+    <?php require_once '_Class/Article.php' ?>
 
     <header>
         <?php require_once '_include/header.php' ?>
