--- connexion.php
@@ -1,15 +1,0 @@
-<?php
-?>
-<div class="connexion-container">
-    <form method="POST" id="formCon">
-        <h2 class="conn-title">Connexion</h2>
-
-        <label class="space conn-text" for="login">Login</label>
-        <input class="input" type="text" name="login" id="login">
-
-        <label class="space conn-text" for="password">Mot de passe</label>
-        <input class="input" type="password" name="password" id="password">
-
-        <input class="button-form" type="submit" value="Connexion" id="connexion">
-    </form>
-</div>
