+++ deconnexion.php
@@ -1,6 +1,6 @@
 <?php
 if(session_status() == PHP_SESSION_NONE){ session_start();}
 
-require_once ('User.php');
+require_once ('_Class/User.php');
 
 User::Disconnect();
