--- inscription.php
@@ -1,19 +1,0 @@
-<?php
-?>
-<div class="inscription-container">
-    <form method="POST" id="formReg">
-
-        <h2 class="register-title">Inscription</h2>
-
-        <label class="space register-text" for="login">Login</label>
-        <input class="input" type="text" name="login" id="login">
-
-        <label class="space register-text" for="password">Mot de passe</label>
-        <input class="input" type="password" name="password" id="password">
-
-        <label class="space register-text" for="confpassword">Confirmation mot de passe</label>
-        <input class="input" type="password" name="confpassword" id="confpassword">
-
-        <input class="button-form" type="submit" value="Inscription" id="inscription">
-    </form>
-</div>
